Move all team-related components to their own directory, cleaned up upgrade messaging
