+++ packages/front-end/components/Settings/Team/RoleDisplay.tsx
@@ -1,0 +1,10 @@
+import { MemberRole } from "back-end/types/organization";
+import { ReactElement } from "react";
+
+export default function RoleDisplay({
+  role,
+}: {
+  role: MemberRole;
+}): ReactElement {
+  return <span className="badge badge-secondary">{role}</span>;
+}
