+++ packages/front-end/components/Settings/Team/RoleSelector.tsx
@@ -1,0 +1,49 @@
+import { MemberRoleWithProjects } from "back-end/types/organization";
+import SingleRoleSelector from "./SingleRoleSelector";
+import RoleUpgradeMessage from "./RoleUpgradeMessage";
+import ProjectRolesSelector from "./ProjectRolesSelector";
+
+export default function RoleSelector({
+  value,
+  setValue,
+  showUpgradeModal,
+  newUser,
+}: {
+  value: MemberRoleWithProjects;
+  setValue: (value: MemberRoleWithProjects) => void;
+  showUpgradeModal: () => void;
+  newUser?: boolean;
+}) {
+  return (
+    <div>
+      <SingleRoleSelector
+        value={{
+          role: value.role,
+          environments: value.environments,
+          limitAccessByEnvironment: value.limitAccessByEnvironment,
+        }}
+        setValue={(newRoleInfo) => {
+          setValue({
+            ...value,
+            ...newRoleInfo,
+          });
+        }}
+        label="Global Role"
+        includeAdminRole={true}
+      />
+      <ProjectRolesSelector
+        projectRoles={value.projectRoles || []}
+        setProjectRoles={(projectRoles) => {
+          setValue({
+            ...value,
+            projectRoles,
+          });
+        }}
+      />
+      <RoleUpgradeMessage
+        showUpgradeModal={showUpgradeModal}
+        newUser={newUser}
+      />
+    </div>
+  );
+}
